+++ bst_mt_lmtx.h
@@ -36,7 +36,7 @@ IN THE SOFTWARE.
 #include "bst_common.h"
 
 typedef struct bst_mt_node {
-    int value;
+    long long int value;
     struct bst_mt_node *left;
     struct bst_mt_node *right;
     pthread_mutex_t mtx;
@@ -47,7 +47,7 @@ typedef struct bst_mt_lmtx {
     pthread_mutex_t mtx;
 } bst_mt_lmtx;
 
-static inline bst_mt_node *bst_mt_node_new(int value) {
+static inline bst_mt_node *bst_mt_node_new(long long int value) {
     bst_mt_node *node = (bst_mt_node *)malloc(sizeof *node);
 
     if (node == NULL) {
@@ -97,7 +97,7 @@ static inline bst_mt_lmtx *bst_mt_lmtx_new() {
     return bst;
 }
 
-static inline int bst_mt_lmtx_add(bst_mt_lmtx *bst, int value) {
+static inline int bst_mt_lmtx_add(bst_mt_lmtx *bst, long long int value) {
     if (bst == NULL) {
         return 1;
     }
@@ -169,7 +169,7 @@ static inline int bst_mt_lmtx_add(bst_mt_lmtx *bst, int value) {
     return 1;
 }
 
-static inline int bst_mt_lmtx_search(bst_mt_lmtx *bst, int value) {
+static inline int bst_mt_lmtx_search(bst_mt_lmtx *bst, long long int value) {
     if (bst == NULL) {
         return 1;
     }
@@ -189,7 +189,7 @@ static inline int bst_mt_lmtx_search(bst_mt_lmtx *bst, int value) {
     return 1;
 }
 
-static inline int bst_mt_lmtx_min(bst_mt_lmtx *bst) {
+static inline long long int bst_mt_lmtx_min(bst_mt_lmtx *bst) {
     if (bst == NULL) {
         return 1;
     }
@@ -207,7 +207,7 @@ static inline int bst_mt_lmtx_min(bst_mt_lmtx *bst) {
     return root->value;
 }
 
-static inline int bst_mt_lmtx_max(bst_mt_lmtx *bst) {
+static inline long long int bst_mt_lmtx_max(bst_mt_lmtx *bst) {
     if (bst == NULL) {
         return 1;
     }
@@ -225,7 +225,7 @@ static inline int bst_mt_lmtx_max(bst_mt_lmtx *bst) {
     return root->value;
 }
 
-static inline int bst_mt_lmtx_node_find_height(bst_mt_node *root) {
+static inline long long int bst_mt_lmtx_node_find_height(bst_mt_node *root) {
     if (root == NULL) {
         return -1;
     }
@@ -234,7 +234,7 @@ static inline int bst_mt_lmtx_node_find_height(bst_mt_node *root) {
                    bst_mt_lmtx_node_find_height(root->right));
 }
 
-static inline int bst_mt_lmtx_height(bst_mt_lmtx *bst) {
+static inline long long int bst_mt_lmtx_height(bst_mt_lmtx *bst) {
     if (bst == NULL) {
         return -1;
     }
@@ -244,7 +244,7 @@ static inline int bst_mt_lmtx_height(bst_mt_lmtx *bst) {
 
 static inline void bst_mt_lmtx_node_traverse_preorder(bst_mt_node *node) {
     if (node != NULL) {
-        printf("%d ", node->value);
+        printf("%lld ", node->value);
         bst_mt_lmtx_node_traverse_preorder(node->left);
         bst_mt_lmtx_node_traverse_preorder(node->right);
     }
@@ -264,7 +264,7 @@ static inline int bst_mt_lmtx_traverse_preorder(bst_mt_lmtx *bst) {
 static inline void bst_mt_lmtx_node_traverse_inorder(bst_mt_node *node) {
     if (node != NULL) {
         bst_mt_lmtx_node_traverse_inorder(node->left);
-        printf("%d ", node->value);
+        printf("%lld ", node->value);
         bst_mt_lmtx_node_traverse_inorder(node->right);
     }
 }
@@ -284,7 +284,7 @@ static inline void bst_mt_lmtx_node_traverse_postorder(bst_mt_node *node) {
     if (node != NULL) {
         bst_mt_lmtx_node_traverse_postorder(node->left);
         bst_mt_lmtx_node_traverse_postorder(node->right);
-        printf("%d ", node->value);
+        printf("%lld ", node->value);
     }
 }
 
@@ -308,7 +308,8 @@ static inline bst_mt_node *lmt_min_node(bst_mt_node *node) {
     return current;
 }
 
-static inline bst_mt_node *lmt_delete_node(bst_mt_node *root, int value) {
+static inline bst_mt_node *lmt_delete_node(bst_mt_node *root,
+                                           long long int value) {
     if (root == NULL) {
         return root;
     }
@@ -337,7 +338,7 @@ static inline bst_mt_node *lmt_delete_node(bst_mt_node *root, int value) {
     return root;
 }
 
-static inline int bst_mt_lmtx_delete(bst_mt_lmtx *bst, int value) {
+static inline int bst_mt_lmtx_delete(bst_mt_lmtx *bst, long long int value) {
     if (bst == NULL) {
         return 1;
     }
@@ -347,8 +348,8 @@ static inline int bst_mt_lmtx_delete(bst_mt_lmtx *bst, int value) {
     return 0;
 }
 
-static inline void lmt_save_inorder(bst_mt_node *node, int *inorder,
-                                    int *index) {
+static inline void lmt_save_inorder(bst_mt_node *node, long long int *inorder,
+                                    long long int *index) {
     if (node == NULL)
         return;
     lmt_save_inorder(node->left, inorder, index);
@@ -356,11 +357,12 @@ static inline void lmt_save_inorder(bst_mt_node *node, int *inorder,
     lmt_save_inorder(node->right, inorder, index);
 }
 
-static inline bst_mt_node *lmt_array_to_bst(int arr[], int start, int end) {
+static inline bst_mt_node *
+lmt_array_to_bst(long long int arr[], long long int start, long long int end) {
     if (start > end)
         return NULL;
 
-    int mid = (start + end) / 2;
+    long long int mid = (start + end) / 2;
     bst_mt_node *node = bst_mt_node_new(arr[mid]);
 
     node->left = lmt_array_to_bst(arr, start, mid - 1);
@@ -369,7 +371,7 @@ static inline bst_mt_node *lmt_array_to_bst(int arr[], int start, int end) {
     return node;
 }
 
-static inline int lmt_node_count(bst_mt_node *root) {
+static inline long long int lmt_node_count(bst_mt_node *root) {
     if (root == NULL) {
         return 0;
     } else {
@@ -377,14 +379,14 @@ static inline int lmt_node_count(bst_mt_node *root) {
     }
 }
 
-static inline int bst_mt_lmtx_width(bst_mt_lmtx *bst) {
+static inline long long int bst_mt_lmtx_width(bst_mt_lmtx *bst) {
     if (bst == NULL) {
         return -1;
     }
 
-    int w = 0;
+    long long int w = 0;
     bst_mt_node **q = malloc(sizeof *q * lmt_node_count(bst->root));
-    int f = 0, r = 0;
+    long long int f = 0, r = 0;
 
     q[r++] = bst->root;
 
@@ -425,8 +427,9 @@ static inline bst_mt_lmtx *bst_mt_lmtx_rebalance(bst_mt_lmtx *bst) {
         return NULL;
     }
 
-    int *inorder = malloc(sizeof(int) * lmt_node_count(bst->root));
-    int index = 0;
+    long long int *inorder =
+        malloc(sizeof(long long int) * lmt_node_count(bst->root));
+    long long int index = 0;
     lmt_save_inorder(bst->root, inorder, &index);
 
     bst_mt_node_free(bst->root);
@@ -441,11 +444,11 @@ static inline void bst_mt_lmtx_print_details(bst_mt_lmtx *bst) {
     if (bst == NULL) {
         return;
     }
-    printf("%d,", lmt_node_count(bst->root));
-    printf("%d,", bst_mt_lmtx_min(bst));
-    printf("%d,", bst_mt_lmtx_max(bst));
-    printf("%d,", bst_mt_lmtx_height(bst));
-    printf("%d,", bst_mt_lmtx_width(bst));
+    printf("%lld,", lmt_node_count(bst->root));
+    printf("%lld,", bst_mt_lmtx_min(bst));
+    printf("%lld,", bst_mt_lmtx_max(bst));
+    printf("%lld,", bst_mt_lmtx_height(bst));
+    printf("%lld,", bst_mt_lmtx_width(bst));
 }
 
 #endif
